High resolution crash

+++ src/Pixel.h
@@ -2,14 +2,15 @@
 
 struct Pixel
 {
-    int r{0};
-    int g{255};
-    int b{0};
-    size_t u{0};
-    size_t v{0};
+    uint8_t r{0};
+    uint8_t g{255};
+    uint8_t b{0};
+    uint16_t u{0};
+    uint16_t v{0};
 };
 
 void operator<<(std::ostream& ostream, const Pixel& pixel)
 {
-    ostream << pixel.r << ' ' << pixel.g << ' ' << pixel.b << '\n';
+    ostream << static_cast<int>(pixel.r) << ' ' << static_cast<int>(pixel.g) << ' '
+            << static_cast<int>(pixel.b) << '\n';
 }
